requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-model-info==2024.11.5
djangorestframework==3.16.1
djangorestframework-api-key==3.1.0
djangorestframework_simplejwt==5.5.1
markdown-it-py==4.0.0
mdurl==0.1.2
networkx==3.5
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.1
rich==14.1.0
sqlparse==0.5.3
